+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 Reworked API Reference organization - now runs through tags instead of through the configuration option `classSections`, done to make it easier to configure classes without having to go to the config, and allowing you to specify a group of a class right in the code, which is more in-line with how Moonwave works.
 
 Added isRoot to Docs.
+
+# 1.3.3
+Fixed several bugs:
+- Base url is now prepended to class links
+- Clean URL support for VitePress